Move common admin-functionality to new home
